### 2019-03-04 #1297 @arfon
:wave: @zmeers - please review the JOSS submission guidelines and draft a paper to support your submission: https://joss.readthedocs.io/en/latest/submitting.html#submitting-your-paper

I'm going to pause this submission until we've heard back from you that you've done this.
