### 2019-03-04 #1297 @zmeers
@arfon -- sorry about that. The paper is actually in the dev branch. Can I submit the paper from the dev branch or does it need to be in the master branch for compilation? I was planning to push the paper to the master branch, once accepted.
