### 2019-03-05 7f491f6
update paper to fix minor errors
