+++ paper/paper.bib
@@ -43,6 +43,8 @@
     note = {R package version 0.3.23},
 }
 
+@Manual{pscl,    title = {{pscl}: Classes and Methods for {R} Developed in the Political Science Computational Laboratory},    author = {Simon Jackman},    organization = {United States Studies Centre, University of Sydney},    address = {Sydney, New South Wales, Australia},    year = {2017},    note = {R package version 1.5.2},    url = {https://github.com/atahk/pscl/},  }
+
 @Book{ggplot2,
     author = {Hadley Wickham},
     title = {ggplot2: Elegant Graphics for Data Analysis},
