### 2019-03-05 b96bb41
update paper to fix minor errors
