+++ paper/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 `ggparliament` is a `R` [@R] package for visualizing legislative chambers. It allows for easy visualization of the party composition of a legislative body or for plotting election results using syntax from the `tidyverse`, a set of packages for cleaning, modelling, and visualizing data [@tidyverse]. The package is one of the first `ggplot2` [@ggplot2] extensions for visualizing political data. It is considerably more detailed than most visualization tools for graphing legislatures in that it is designed to handle many different legislatures irrespective of size, number of parties, or characteristics of legislatures.  Legislature visualizations are also known as parliament plots or parliament charts.
 
-Visualizing legislatures can be done in a variety of formats - usually in a JavaScript framework. `ggparliament` seeks to implement the flexibility from JavaScript in `R`. `R` users will find they can use a similar syntax to the popular `ggplot2` package to generate a visualization of a legislative body. In particular, `ggparliament` is a `R` package because `R` is a statistical computing language that many quantitative political scientists use in research.[^See `pscl` [@pscl], `Amelia` [@Amelia], `MatchIt` [@MatchIt], and `margins` [@margins] for an incomplete list of `R` packages created for political science research.] 
+Visualizing legislatures can be done in a variety of formats - usually in a JavaScript framework. `ggparliament` seeks to implement the flexibility from JavaScript in `R`. `R` users will find they can use a similar syntax to the popular `ggplot2` package to generate a visualization of a legislative body. In particular, `ggparliament` is a `R` package because `R` is a statistical computing language that many quantitative political scientists use in research.[^See `pscl` [@pscl], `Amelia` [@Amelia], `MatchIt` [@MatchIt], and `margins` [@margins] for an incomplete list of `R` packages created primarily for political science research.] 
 
 `ggparliament` is useful research tool for a variety of social science disciplines, including quantitative political science. It is particularly beneficial for political scientists who research political institutions, such as electoral systems, party politics, or legislative politics. `ggparliament` provides several layouts, representing different legislative chambers e.g. the United Kingdom's House of Commons, Australia's horseshoe-shaped parliament, or the widely-used semicircle legislative chamber. 
 
