### 2019-03-05 #1297 @zmeers
Some minor issues with the BibTex...
