### 2019-03-05 dcf02e4
update paper to remove footnote
