### 2019-03-05 #1297 @zmeers
The foot note did not compile. I will change it and recompile the paper.
