### 2019-03-11 #1313 @whedon
[REVIEW]: ggparliament: A ggplot2 extension for parliament plots in R
**Submitting author:** @zmeers (<a href="http://orcid.org/0000-0001-8045-6531">Zoe Meers</a>)
**Repository:** <a href="https://github.com/RobWHickman/ggparliament/" target ="_blank">https://github.com/RobWHickman/ggparliament/</a>
**Version:** v.2.1.0

- [x] **Version:** v.2.1.0
