### 2019-04-08 #1313 @andrewheiss
Sorry for the delay! Here's my review:

---

**ggparliament** is a fantastic new package that extends **ggplot2** with custom geometries for political data. It includes a dataset of a handful of parliamentary elections and provides an easy-to-use function for creating your own legislative data. The `geom_*()` and `theme_ggparliament()` functions it provides are well documented and easy to follow, and the package's documentation in general is well done. I only came across a few issues, noted below.

## Documentation

The README runs without any problems and is easy to follow.

It might be helpful to include documentation (perhaps in the README) about what is included in the sample data. This could be as simple as adding a chunk like this to the README:

```r
election_data %>%
  distinct(year, country)

# or

election_data %>%
  distinct(year, country, house)
```

The vignettes are fantastically written, are easy to follow, and they all run correctly. They should maybe be included in a list in the README so that people who come across the package through GitHub know that they exist. (Though GitHub doesn't render R Markdown nicely, unfortunately; but if the documentation is ever published online through **pkgdown** or something similar, users would be able to get to the vignettes more easily)

In the `facet_parliament_5.Rmd` vignette, there's a warning caused by line 97 because of a (vestigial?) `type = "horseshoe"` in `ggplot(aes(...))`. The `parliament_data()` function seems to take care of chart types now, so the type aesthetic doesn't do anything in `ggplot()` anymore. Removing it from `ggplot()` fixes the warning.

## Testing

There are two tests related to the structure of data created with `parliament_data()` and they pass. Adding other tests would be difficult, since they deal with visual output. **ggplot2** includes [detailed tests of its figures](https://github.com/tidyverse/ggplot2/tree/master/tests/testthat) by delving into the resulting grid objects, but that's probably not necessary here.

## Contributions

It might be helpful to have some community guidelines in the package too, such as a CONTRIBUTING.md file (perhaps modeled after something like [this](https://github.com/ropensci/nomisr/blob/master/CONTRIBUTING.md) or [this](https://github.com/tidyverse/ggplot2/blob/master/CONTRIBUTING.md)) and a CONDUCT.md file (like [this](https://github.com/ropensci/qualtRics/blob/master/CONDUCT.md) or [this](https://github.com/tidyverse/readr/blob/master/CONDUCT.md))

Creating an RStudio project `.Rproj` file could be helpful for encouraging package contributions, since it would enable RStudio's package building interface.

## DOIs

`paper/paper.bib` file should include DOIs where possible—right now, none of the entries have a DOI

----

That's all I have—I'm excited for this package!
