### 2019-04-18 c378625
addressing reviewer comments from JOSS -- add list of of vignettes to README and remove type = "horseshoe" from one of the vignettes.
