+++ R/geom_parliament_seats.R
@@ -15,8 +15,8 @@
 #' @examples
 #' data <- election_data[
 #'   election_data$country == "USA" &
-#'   election_data$house == "Representatives" &
-#'   election_data$year == "2016",
+#'     election_data$house == "Representatives" &
+#'     election_data$year == "2016",
 #' ]
 #' usa_data <- parliament_data(
 #'   election_data = data,
@@ -58,33 +58,33 @@ geom_parliament_seats <- function(mapping = NULL,
 #' @usage NULL
 #' @export
 GeomParliamentSeats <- ggplot2::ggproto("GeomParliamentSeats", ggplot2::Geom,
-                                        required_aes = c("x", "y", "colour"),
-                                        non_missing_aes = c("shape", "size"),
-                                        default_aes = ggplot2::aes(
-                                          shape = 19,
-                                          colour = "black",
-                                          size = 3.5,
-                                          fill = NA,
-                                          alpha = NA,
-                                          stroke = 1
-                                        ),
-                                        
-                                        draw_panel = function(data, panel_params, coord, na.rm = FALSE) {
-                                          coords <- coord$transform(data, panel_params)
-                                          ggplot2:::ggname(
-                                            "geom_parliament_seats",
-                                            grid::pointsGrob(
-                                              coords$x, coords$y,
-                                              pch = coords$shape,
-                                              gp = grid::gpar(
-                                                col = alpha(coords$colour, coords$alpha),
-                                                fill = alpha(coords$fill, coords$alpha),
-                                                fontsize = coords$size * ggplot2::.pt + coords$stroke * .stroke,
-                                                lwd = coords$stroke * .stroke
-                                              )
-                                            )
-                                          )
-                                        },
-                                        
-                                        draw_key = ggplot2::draw_key_point
+  required_aes = c("x", "y", "colour"),
+  non_missing_aes = c("shape", "size"),
+  default_aes = ggplot2::aes(
+    shape = 19,
+    colour = "black",
+    size = 3.5,
+    fill = NA,
+    alpha = NA,
+    stroke = 1
+  ),
+
+  draw_panel = function(data, panel_params, coord, na.rm = FALSE) {
+    coords <- coord$transform(data, panel_params)
+    ggplot2:::ggname(
+      "geom_parliament_seats",
+      grid::pointsGrob(
+        coords$x, coords$y,
+        pch = coords$shape,
+        gp = grid::gpar(
+          col = alpha(coords$colour, coords$alpha),
+          fill = alpha(coords$fill, coords$alpha),
+          fontsize = coords$size * ggplot2::.pt + coords$stroke * .stroke,
+          lwd = coords$stroke * .stroke
+        )
+      )
+    )
+  },
+
+  draw_key = ggplot2::draw_key_point
 )
