+++ R/geom_emphasize_parliamentarians.R
@@ -6,8 +6,8 @@
 #' @examples
 #' data <- election_data[
 #'   election_data$country == "USA" &
-#'   election_data$house == "Representatives" &
-#'   election_data$year == "2016",
+#'     election_data$house == "Representatives" &
+#'     election_data$year == "2016",
 #' ]
 #' usa_data <- parliament_data(
 #'   election_data = data,
@@ -15,13 +15,13 @@
 #'   party_seats = data$seats,
 #'   parl_rows = 8
 #' )
-#'
+#' 
 #' women_in_congress <- c(1, 0, 0, 1)
 #' number_of_women <- c(23, 218, 133, 61)
-#'
+#' 
 #' usa_data$women <- rep(women_in_congress, number_of_women)
-#'
-#' ggplot2::ggplot(usa_data, ggplot2::aes(x, y, colour=party_long)) +
+#' 
+#' ggplot2::ggplot(usa_data, ggplot2::aes(x, y, colour = party_long)) +
 #'   geom_parliament_seats() +
 #'   geom_emphasize_parliamentarians(women == 1) +
 #'   theme_ggparliament(legend = FALSE) +
@@ -33,12 +33,15 @@
 #' @export
 #' @importFrom ggplot2 ggplot_add
 
-geom_emphasize_parliamentarians <- function(expr,  size = 3, shape = 19, stroke = 1.8) {
-  structure(list(expr = rlang::enquo(expr),
-                 size = size,
-                 shape = shape,
-                 stroke = stroke), 
-            class = "emphMPs")
+geom_emphasize_parliamentarians <- function(expr, size = 3, shape = 19, stroke = 1.8) {
+  structure(list(
+    expr = rlang::enquo(expr),
+    size = size,
+    shape = shape,
+    stroke = stroke
+  ),
+  class = "emphMPs"
+  )
 }
 
 
