+++ man/draw_majoritythreshold.Rd
@@ -29,8 +29,8 @@ Draw majority threshold
 \examples{
 data <- election_data[
   election_data$country == "USA" &
-  election_data$house == "Representatives" &
-  election_data$year == "2016",
+    election_data$house == "Representatives" &
+    election_data$year == "2016",
 ]
 usa_data <- parliament_data(
   election_data = data,
@@ -43,7 +43,7 @@ ggplot2::ggplot(usa_data, ggplot2::aes(x, y, colour = party_long)) +
   draw_majoritythreshold(
     n = 218,
     label = TRUE,
-    type = 'semicircle'
+    type = "semicircle"
   ) +
   theme_ggparliament()
 }

[…]
