+++ CONTRIBUTING.md
@@ -1,0 +1,27 @@
+# Contributing to `ggparliament`
+
+
+Thank you for any contributions. We welcome contributions from all, regardless of your programming experience. 
+
+## Types of contributions 
+
+We would love to hear from anyone who is willing to provide their time and help.
+
+Here is a non-exhaustive list of things that we might need help with:
+1. Election data
+2. Adding additional information about legislatures or electoral systems
+3. Extending interactive `ggparliament` objects
+4. Writing tutorials or vignettes
+5. Fix bugs
+
+
+## How to contribute code or data
+
+1. Fork this repository
+2. Make your changes and commit to your GitHub fork. 
+3. Open an issue here and submit a pull request. 
+
+
+
+
+
