+++ README.Rmd
@@ -59,15 +59,13 @@ If you have any issues, please note the problem and inform us!
 
 ## Election data
 
-`ggparliament` provides electiond data from the following countries. 
+`ggparliament` provides election data from the following countries. 
 
 ```{r}
 election_data %>% 
   distinct(year, country, house) %>% 
   arrange(country, year)
 ```
-
-
 
 We also provide the following vignettes for further explanation:
 
