+++ README.Rmd
@@ -16,17 +16,17 @@ library(tidyr)
 library(ggplot2)
 
 
-# source("R/parliament_data.R")
-# source("R/geom_parliament_seats.R")
-# source("R/geom_highlight_government.R")
-# source("R/geom_emphasize_parliamentarians.R")
-# source("R/helper_funcs.R")
-# source("R/draw_majoritythreshold.R")
-# source("R/draw_partylabels.R")
-# source("R/draw_majoritythreshold.R")
-# source("R/draw_totalseats.R")
-# source("R/theme_ggparliament.R")
-# source("R/geom_parliament_bar.R")
+source("R/parliament_data.R")
+source("R/geom_parliament_seats.R")
+source("R/geom_highlight_government.R")
+source("R/geom_emphasize_parliamentarians.R")
+source("R/helper_funcs.R")
+source("R/draw_majoritythreshold.R")
+source("R/draw_partylabels.R")
+source("R/draw_majoritythreshold.R")
+source("R/draw_totalseats.R")
+source("R/theme_ggparliament.R")
+source("R/geom_parliament_bar.R")
 
 ```
 
