+++ README.md
@@ -9,21 +9,266 @@ pagetitle: README
 
 
 
+# Status
+[![Build Status](https://travis-ci.org/RobWHickman/ggparliament.png)](https://travis-ci.org/RobWHickman/ggparliament)
+![CRAN Status](https://www.r-pkg.org/badges/version/ggparliament)
+![Downloads](https://cranlogs.r-pkg.org/badges/grand-total/ggparliament)
+<img src = "man/figures/HexSticker.png" align = "right" width = "200"/>
+
+# ggparliament: Parliament plots
+
+
+This package attempts to implement "parliament plots" - visual representations of the composition of legislatures that display seats colour-coded by party. The input is a data frame containing one row per party, with columns representing party name/label and number of seats, respectively.
+
+This `R` package is a `ggplot2` extension and is now on CRAN. Please install the stable version in `R` by running:
+
+
+```r
+install.packages("ggparliament")
+```
+
+To install the package from source:
+
+```r
+devtools::install_github("robwhickman/ggparliament")
+```
+
+Inspiration from this package comes from: [parliamentdiagram](https://github.com/slashme/parliamentdiagram), which
+is used on Wikipedia, [parliament-svg](https://github.com/juliuste/parliament-svg), which is a javascript clone, and [a discussion on StackOverflow](http://stackoverflow.com/questions/42729174/creating-a-half-donut-or-parliamentary-seating-chart), which provided some of the code for part for the "arc" representations used in this package.
+
+
+If you have any issues, please note the problem and inform us!
+
+
+## Election data
+
+`ggparliament` provides election data from the following countries. 
+
+
+```r
+election_data %>% 
+  distinct(year, country, house) %>% 
+  arrange(country, year)
+```
+
+```
+##    year   country           house
+## 1  2010 Australia Representatives
+## 2  2010 Australia          Senate
+## 3  2013 Australia Representatives
+## 4  2013 Australia          Senate
+## 5  2016 Australia Representatives
+## 6  2016 Australia          Senate
+## 7  1990   Germany       Bundestag
+## 8  1994   Germany       Bundestag
+## 9  1998   Germany       Bundestag
+## 10 2002   Germany       Bundestag
+## 11 2005   Germany       Bundestag
+## 12 2009   Germany       Bundestag
+## 13 2013   Germany       Bundestag
+## 14 2017   Germany       Bundestag
+## 15 2007    Russia            Duma
+## 16 2011    Russia            Duma
+## 17 2016    Russia            Duma
+## 18 2010        UK         Commons
+## 19 2015        UK         Commons
+## 20 2017        UK         Commons
+## 21 2012       USA          Senate
+## 22 2012       USA Representatives
+## 23 2014       USA          Senate
+## 24 2014       USA Representatives
+## 25 2016       USA          Senate
+## 26 2016       USA Representatives
+```
+
+We also provide the following vignettes for further explanation:
+
+1. Basic parliament plots
+2. Labelling parties
+3. Drawing the majority threshold line
+4. Highlighting parties in power
+5. Faceting legislatures
+6. Emphasizing certain seats
+7. Visualizaing overhang seats in MMP electoral systems
+8. Arranging seat order in ggparliament plots.
+
+Quick `ggparliament` examples can be viewed below.
+
+## Semicircle parliament
+
+### EU, France, United States, and so on...
+
+
+### Plot of US House of Representatives
 
 
 
+```r
+#filter the election data for the most recent US House of Representatives
+us_house <- election_data %>%
[…]
