+++ README.Rmd
@@ -10,7 +10,7 @@ pagetitle: README
 ```{r, include=FALSE}
 knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
 
-library(ggparliament)
+#library(ggparliament)
 library(dplyr)
 library(tidyr)
 library(ggplot2)
@@ -27,7 +27,7 @@ source("R/draw_majoritythreshold.R")
 source("R/draw_totalseats.R")
 source("R/theme_ggparliament.R")
 source("R/geom_parliament_bar.R")
-
+load("data/election_data.rda")
 ```
 
 # Status
@@ -36,7 +36,7 @@ source("R/geom_parliament_bar.R")
 ![Downloads](https://cranlogs.r-pkg.org/badges/grand-total/ggparliament)
 <img src = "man/figures/HexSticker.png" align = "right" width = "200"/>
 
-# ggparliament: Parliament plots
+# `ggparliament`: Parliament plots
 
 
 This package attempts to implement "parliament plots" - visual representations of the composition of legislatures that display seats colour-coded by party. The input is a data frame containing one row per party, with columns representing party name/label and number of seats, respectively.
