### 2019-04-18 #1313 @zmeers
@andrewheiss, thank you!

@cMadan, I have updated the README file, removed the error from one of the vignettes, and have added a CONTRIBUTING.md file. I have also included DOIs where possible in the bibliography. Thanks.
