### 2019-04-22 #1313 @cMadan
@zmeers, everything looks good to me! Can you create a Github release corresponding to the current repo and tell me the new version number that now incorporates the recent changes? You also need to deposit the code somewhere archival (e.g., Zenodo or figshare) and tell me the DOI? Thanks!
