### 2019-04-25 #1313 @zmeers
@cMaden, Sure, it's all done. The updated version of the R package is v.2.1.0 in the dev branch. The DOI for Zenodo is 10.5281/zenodo.2650709. Thanks!
