### 2019-04-25 #1313 @cMadan
@openjournals/joss-eics, I think we're all set to accept here!
