### 2019-04-25 #1313 @danielskatz
@zmeers - please add DOIs for references that have them, such as the journal papers.  See the [example files](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) if needed
