### 2019-04-25 #1313 @arfon
> 👋 @arfon - any idea what's going on? Either check references isn't completing, or it's very slow. Also note that it didn't complete when accept was run.

I'm guessing that this line might be causing failures: https://github.com/RobWHickman/ggparliament/blob/master/paper/paper.bib#L46 @zmeers - any chance you could take a look at cleaning up this entry?
