### 2019-04-26 #1313 @zmeers
Hey @arfon, I believe all of the issues have been fixed in the dev branch. I've added DOIs for all articles or entries that have DOI as well. Unfortunately, a few entries do not have DOIs. You will need to compile the paper from the dev branch though, I hope that's not a problem!
