### 2019-04-26 #1313 @danielskatz
@zmeers - is there a reason you can't put these changes in the master branch?
