### 2019-04-26 #1313 @zmeers
@arfon, @danielskatz, it's now in the master branch! Sorry for any confusion.
